Update substrate, evm and ethereum versions (#275)

+++ frame/evm/precompile/modexp/src/lib.rs
@@ -43,12 +43,12 @@ pub struct Modexp;
 //       see: https://eips.ethereum.org/EIPS/eip-198
 
 impl LinearCostPrecompile for Modexp {
-	const BASE: usize = 15;
-	const WORD: usize = 3;
+	const BASE: u64 = 15;
+	const WORD: u64 = 3;
 
 	fn execute(
 		input: &[u8],
-		_: usize,
+		_: u64,
 	) -> core::result::Result<(ExitSucceed, Vec<u8>), ExitError> {
 		if input.len() < 96 {
 			return Err(ExitError::Other("input must contain at least 96 bytes".into()));
@@ -137,9 +137,9 @@ mod tests {
 
 	#[test]
 	fn test_empty_input() -> std::result::Result<(), ExitError> {
-
 		let input: [u8; 0] = [];
-		let cost: usize = 1;
+
+		let cost: u64 = 1;
 
 		match Modexp::execute(&input, cost) {
 			Ok((_, _)) => {
@@ -154,14 +154,13 @@ mod tests {
 
 	#[test]
 	fn test_insufficient_input() -> std::result::Result<(), ExitError> {
-
 		let input = hex::decode(
 			"0000000000000000000000000000000000000000000000000000000000000001\
 			0000000000000000000000000000000000000000000000000000000000000001\
 			0000000000000000000000000000000000000000000000000000000000000001")
 			.expect("Decode failed");
 
-		let cost: usize = 1;
+		let cost: u64 = 1;
 
 		match Modexp::execute(&input, cost) {
 			Ok((_, _)) => {
@@ -183,7 +182,7 @@ mod tests {
 			0000000000000000000000000000000000000000000000000000000000000001")
 			.expect("Decode failed");
 
-		let cost: usize = 1;
+		let cost: u64 = 1;
 
 		match Modexp::execute(&input, cost) {
 			Ok((_, _)) => {
@@ -209,7 +208,7 @@ mod tests {
 
 		// 3 ^ 5 % 7 == 5
 
-		let cost: usize = 1;
+		let cost: u64 = 1;
 
 		match Modexp::execute(&input, cost) {
 			Ok((_, output)) => {
@@ -238,7 +237,7 @@ mod tests {
 
 		// 59999 ^ 21 % 14452 = 10055
 
-		let cost: usize = 1;
+		let cost: u64 = 1;
 
 		match Modexp::execute(&input, cost) {
 			Ok((_, output)) => {
@@ -255,7 +254,6 @@ mod tests {
 
 	#[test]
 	fn test_large_computation() {
-
 		let input = hex::decode(
 			"0000000000000000000000000000000000000000000000000000000000000001\
 			0000000000000000000000000000000000000000000000000000000000000020\
@@ -265,7 +263,7 @@ mod tests {
 			fffffffffffffffffffffffffffffffffffffffffffffffffffffffefffffc2f")
 			.expect("Decode failed");
 
-		let cost: usize = 1;
+		let cost: u64 = 1;
 
 		match Modexp::execute(&input, cost) {
 			Ok((_, output)) => {
